fix: separate default values from interface
Due to the interface setting up both the types and the default values, when in use with TS a type mismatch error occurs where any of the attributes cannot be set, without an error appearing. Due to that, separating types in the interface from the default values should fix said issue.